Fix inline pagelet processing in Rails 2.3
Although rails 2.3 follows the Rack specification, it
doesn't return a simple string in the content part of
the rack response, instead it returns an instance of
ActionController::Response.

Just trying to #to_s this into the template results in
the template.

Fix by always calling #body.

There doesn't seem to be a clean way to test this at the
moment, as the test app is for Rails 2.2, and it seems
overkill to create a whole new test app for this case.